refactor(value-node): rename DataNode folder to ValueNode
